requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\newap\Desktop\NotesManager-Beta\main.py: 5
# C:\Users\newap\Desktop\NotesManager-Beta\notes\core\categorybase.py: 3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\core\categorycolor.py: 4
# C:\Users\newap\Desktop\NotesManager-Beta\notes\core\config.py: 8,10,11
# C:\Users\newap\Desktop\NotesManager-Beta\notes\core\noteattachment.py: 3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\core\notebooksettings.py: 3,4,5
# C:\Users\newap\Desktop\NotesManager-Beta\notes\core\notecategory.py: 1
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\MainWindow.py: 6
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\Ui_MainWindow.py: 10,11
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\cetegorybaseview.py: 1,2,3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\collapsiblebox.py: 2
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\ext\find.py: 3,4
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\ext\table.py: 3,7
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\ext\wordcount.py: 3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\linkadddialog.py: 1,2,3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\nameeditdialog.py: 3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\noteattachmentsview.py: 4,5,6
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\notebookfilterdialog.py: 1,2,3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\notebookfilteredwidget.py: 1
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\notebookpreviewtreewidget.py: 3,4,5
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\notebooktreewidget.py: 3,4,5
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\notecategoryedit.py: 1,2,3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\notecontentedit.py: 1,2,3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\noteentrieedit.py: 3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\notehistoryview.py: 1,2,3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\noteselectdialog.py: 1,2,3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\noteviewmanager.py: 6
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\settingsdialog.py: 3,4,5
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\tabwidget.py: 1
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\tagbaseview.py: 3,4,5
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\tagedit.py: 3,4,5
PyQt5_sip == 12.7.0

# C:\Users\newap\Desktop\NotesManager-Beta\main.py: 4
QDarkStyle == 2.8

# C:\Users\newap\Desktop\NotesManager-Beta\notes\core\notebook.py: 9
html2text == 2020.1.16

# C:\Users\newap\Desktop\NotesManager-Beta\main.py: 3
# C:\Users\newap\Desktop\NotesManager-Beta\notes\core\note.py: 5
# C:\Users\newap\Desktop\NotesManager-Beta\notes\core\notebook.py: 10
# C:\Users\newap\Desktop\NotesManager-Beta\notes\ui\MainWindow.py: 4
jsonpickle == 1.2
